requirements.txt
aiosqlite
fastapi
httpx
pytest
python-dotenv
sqlalchemy
sqlmodel
uvicorn
